+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup, find_packages
 
 setup(
-    name='tomopyui',
+    name='topy',
     version=open('VERSION').read().strip(),
     #version=__version__,
     author='Matthias Vogelgesang',
@@ -9,8 +9,8 @@ setup(
     url='https://github.com/ufo-kit',
     packages=find_packages(),
     package_data={'':['gui.ui']},
-    scripts=['bin/tomopyui'],
-    description='Basic tomographic reconstruction GUI',
+    scripts=['bin/topy'],
+    description='tofu for tomopy',
     install_requires=['pyqtgraph'],
     zip_safe=False,
 )
