+++ topy/util.py
